+++ Satyamevjayate_API/src/main/java/com/satyamevjayate/api/repo/FormVerificationRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.math.BigInteger;
 
-public interface FormVerificationRepository extends JpaRepository<Formverification, BigInteger> {
+public interface FormVerificationRepository extends JpaRepository<Formverification, Long> {
 }
