+++ Satyamevjayate_API/src/main/java/com/satyamevjayate/api/repo/PoliceRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.satyamevjayate.api.dto.PoliceRequest;
 import com.satyamevjayate.api.entity.Police;
 
-public interface PoliceRepository extends JpaRepository<Police,BigInteger> {
+public interface PoliceRepository extends JpaRepository<Police,Long> {
